updated ui, shifted profile to DashboardFragment.kt, clean code
