+++ app/src/main/AndroidManifest.xml
@@ -17,12 +17,6 @@
         android:theme="@style/Theme.Doubtless"
         tools:targetApi="31">
         <activity
-            android:name=".ProfileActivity"
-            android:exported="true"
-            android:screenOrientation="portrait">
-
-        </activity>
-        <activity
             android:name=".LoginActivity"
             android:exported="true"
             android:screenOrientation="portrait">
@@ -41,13 +35,6 @@
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
-
-        <activity android:name=".LoginActivity"
-            android:exported="false"
-            android:screenOrientation="portrait"/>
-        <activity android:name=".ProfileActivity"
-            android:exported="false"
-            android:screenOrientation="portrait"/>
         <meta-data
             android:name="preloaded_fonts"
             android:resource="@array/preloaded_fonts" />
