--- app/src/main/java/com/doubtless/doubtless/TestActivity.kt
@@ -1,11 +1,0 @@
-package com.doubtless.doubtless
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-class TestActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_test)
-    }
-}
