+++ app/src/main/java/com/doubtless/doubtless/screens/dashboard/DashboardFragment.kt
@@ -1,14 +1,21 @@
 package com.doubtless.doubtless.screens.dashboard
 
+import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
+import com.doubtless.doubtless.LoginActivity
+import com.doubtless.doubtless.R
 import com.doubtless.doubtless.databinding.FragmentDashboardBinding
+import com.google.android.gms.auth.api.signin.GoogleSignIn
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions
+import com.google.firebase.auth.FirebaseAuth
 
 class DashboardFragment : Fragment() {
+    private lateinit var mAuth: FirebaseAuth
 
     private var _binding: FragmentDashboardBinding? = null
 
@@ -25,6 +32,22 @@ class DashboardFragment : Fragment() {
             ViewModelProvider(this).get(DashboardViewModel::class.java)
 
         _binding = FragmentDashboardBinding.inflate(inflater, container, false)
+
+        mAuth = FirebaseAuth.getInstance()
+
+        binding.tvName.text = mAuth.currentUser?.displayName
+        binding.tvUserEmail.text = mAuth.currentUser?.email
+
+        binding.btnSignout.setOnClickListener {
+            val gso = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+                .requestIdToken(getString(R.string.default_web_client_id))
+                .requestEmail()
+                .build()
+
+            val googleSignInClient = GoogleSignIn.getClient(requireContext(),gso)
+            googleSignInClient.signOut()
+            startActivity(Intent(context, LoginActivity::class.java))
+        }
 
         return binding.root
     }
